correcciones de clase

+++ product/destroy.php
@@ -1,9 +1,15 @@
 <?php
+
+ini_set('display_errors', 1);
+error_reporting(E_ALL);
+
 session_start();
 if(!isset($_SESSION['user'])) {
-    header('.');
+    header('Location:.');
     exit;
 }
+$user = $_SESSION['user'];
+
 try {
     $connection = new \PDO(
       'mysql:host=localhost;dbname=productdatabase',
@@ -14,26 +20,36 @@ try {
         PDO::MYSQL_ATTR_INIT_COMMAND => 'set names utf8')
     );
 } catch(PDOException $e) {
-    echo 'no connection';
+    header('Location:..');
     exit;
 }
+
 if(isset($_GET['id'])) {
     $id = $_GET['id'];
 } else {
-    echo 'no id';
+    $url = '.?op=destroyproduct&result=noid';
+    header('Location: ' . $url);
     exit;
 }
+
+if(($user === 'even' && $id % 2 != 0) ||
+    ($user === 'odd' && $id % 2 == 0)) {
+    header('Location: .?op=destroyproduct&result=evenodd');
+    exit;
+}
+
 $sql = 'delete from product where id = :id';
 $sentence = $connection->prepare($sql);
 $parameters = ['id' => $id];
 foreach($parameters as $nombreParametro => $valorParametro) {
     $sentence->bindValue($nombreParametro, $valorParametro);
 }
-if(!$sentence->execute()){
-    echo 'no sql';
-    exit;
+try {
+    $sentence->execute();
+    $resultado = $sentence->rowCount();
+} catch(PDOException $e) {
+    $resultado = 0;
 }
-$resultado = $sentence->rowCount();
 $connection = null;
 $url = '.?op=deleteproduct&result=' . $resultado;
 header('Location: ' . $url);
